Decouple postsGetTrendingPosts saga;
