+++ src/services/__mocks__/Query.js
@@ -1,0 +1,1 @@
+export const apiRequest = jest.fn()
